+++ projects/ngx-error-emperor/src/lib/classes/assigned-error/assigned-error.class.ts
@@ -1,8 +1,10 @@
 import {ErrorContext} from '../error-context/error-context.class';
 
-export class AssignedError {
+export class AssignedError extends Error {
     public constructor(
         public readonly error: unknown,
         public readonly context: ErrorContext
-    ) {}
+    ) {
+        super('AssignedError');
+    }
 }