+++ projects/ngx-error-emperor/src/lib/injection-tokens/console.injection-token.ts
@@ -3,6 +3,6 @@ import {InjectionToken} from '@angular/core';
 export const CONSOLE: InjectionToken<Console> = new InjectionToken('Console', {
     providedIn: 'any',
     factory(): Console {
-        return globalThis?.console ?? window?.console ?? console;
+        return globalThis.console;
     }
 });