+++ projects/ngx-error-emperor/src/lib/injection-tokens/form-container.injection-token.ts
@@ -1,0 +1,7 @@
+import {InjectionToken} from '@angular/core';
+
+import type {FormContainer} from '../types/form-container.type';
+
+export const FORM_CONTAINER: InjectionToken<FormContainer> = new InjectionToken(
+    'FORM_CONTAINER'
+);