+++ projects/ngx-error-emperor/src/lib/services/default-error-handler/default.error-handler.ts
@@ -1,7 +1,7 @@
 import {ErrorHandler, Inject, Injectable} from '@angular/core';
 
 import {AssignedError} from '../../classes/assigned-error/assigned-error.class';
-import {CONSOLE} from '../../injection-tokens/console-injection-token.value';
+import {CONSOLE} from '../../injection-tokens/console.injection-token';
 
 @Injectable()
 export class DefaultErrorHandler implements ErrorHandler {