+++ projects/ngx-error-emperor/src/lib/types/form-container.type.ts
@@ -1,0 +1,5 @@
+import type {Observable} from 'rxjs';
+
+export interface FormContainer {
+    submit(): Promise<void> | Observable<void> | void;
+}