+++ projects/ngx-error-emperor/src/lib/types/form-error-options.type.ts
@@ -1,0 +1,8 @@
+import type {ValidationErrors} from '@angular/forms';
+
+export interface FormErrorOptions {
+    readonly originalError: unknown;
+
+    readonly controlErrors?: Readonly<Record<string, ValidationErrors | null>>;
+    readonly generalError?: unknown;
+}