+++ projects/ngx-error-emperor/src/lib/values/no-error.value.ts
@@ -1,0 +1,1 @@
+export const noError: symbol = Symbol('No error');