+++ projects/ngx-error-emperor/src/public-api.ts
@@ -1,13 +1,21 @@
 export * from './lib/classes/assigned-error/assigned-error.class';
 export * from './lib/classes/error-context/error-context.class';
 export * from './lib/classes/error-context-token/error-context-token.class';
+export * from './lib/classes/form-error/form-error.class';
+
+export * from './lib/components/form/form.component';
 
 export * from './lib/functions/assign-error/assign-error.function';
 
 export * from './lib/types/error-emperor-module-options.type';
 export * from './lib/types/error-interceptor.type';
+export * from './lib/types/form-container.type';
+export * from './lib/types/form-error-options.type';
 
 export * from './lib/injection-tokens/error-handler.injection-token';
 export * from './lib/injection-tokens/error-interceptors.injection-token';
+export * from './lib/injection-tokens/form-container.injection-token';
+
+export * from './lib/values/no-error.value';
 
 export * from './lib/ngx-error-emperor.module';